pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai"
version = "0.0.1"
description = "Breast Concer ML Project"

requires-python = ">= 3.8, < 3.11"

dependencies = [
    "torch ~= 1.13",
    "torchmetrics ~= 0.11",
    "pytorch-lightning ~= 1.9",
    "monai ~= 1.1",
    "Pillow ~= 9.5",
    "type ~= 0.9"
]

[project.optional-dependencies]
lint = [
    "black ~= 22.6",
    "autoflake ~= 1.4",
    "flake8 ~= 4.0",
    "pyproject-flake8 ~= 0.0.1a4",
    "isort ~= 5.10",
    "pre-commit ~= 2.20",
]
dev = ["ai[lint]"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line_length = 120

[tool.isort]
profile = "black"
line_length = 120
order_by_type = true
lines_between_types = 1
combine_as_imports = true
force_grid_wrap = 2
